test: lift ModelStub out
